moved functions to firebase folder
