+++ components/elements/AddCityForm.js
@@ -8,52 +8,26 @@ import _Button from "./_Button";
 import { getAuth } from "firebase/auth";
 import { db } from "../../firebaseConfig";
 
-import { arrayUnion, doc, getDoc, updateDoc } from "firebase/firestore";
+import { arrayUnion, doc, updateDoc } from "firebase/firestore";
+
+import { fetchUsersCities, addUserCity } from "../../firebase/users";
 
 const AddCityForm = () => {
   const [city, setCity] = useState(null);
   const [usersCities, setUsersCities] = useState([]);
-  const fetchUsersCities = async () => {
-    try {
-      const auth = getAuth();
-      const userId = auth.currentUser?.uid;
-      const userRef = doc(db, "users", userId);
-      const userSnap = await getDoc(userRef);
-      console.log("cities:", userSnap.data().cities);
-      if (userSnap.exists()) {
-        setUsersCities(() => {
-          const cities = userSnap.data().cities || [userSnap.data().location];
-          return cities.map((city, i) => ({
-            city: city,
-            id: Date.now() + i + Math.random(),
-          }));
-        });
-      } else {
-        throw new Error("No user found");
-      }
-    } catch (error) {
-      console.error(error);
-    }
-  };
+
   const handleAddCity = async () => {
     try {
-      const auth = getAuth();
-      const userId = auth.currentUser?.uid;
-      const userRef = doc(db, "users", userId);
-      if (!city) throw new Error("No city selected");
-      if (usersCities.find((userCity) => userCity.city === city))
-        throw new Error("City already added");
-      await updateDoc(userRef, {
-        cities: arrayUnion(city),
-      });
-      await fetchUsersCities();
+      addUserCity(city, usersCities);
+      setCity(null);
+      fetchUsersCities(setUsersCities);
     } catch (error) {
       console.error(error);
     }
   };
 
   useEffect(() => {
-    fetchUsersCities();
+    fetchUsersCities(setUsersCities);
   }, []);
 
   return (
